% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbPalette.R
\name{cbPalette}
\alias{cbPalette}
\title{A Colorblind-friendly Palette Function}
\usage{
cbPalette(n)
}
\arguments{
\item{n}{Number of colors to return}
}
\description{
This palette can be used all over
}
\examples{
cbPalette(3)
}
