% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbookPalette.R
\docType{data}
\name{colorbook}
\alias{colorbook}
\title{Color Book Color Schemes}
\format{
An object of class \code{list} of length 6.
}
\usage{
colorbook
}
\description{
https://www.colorbook.io/colorschemes
}
\keyword{datasets}
