% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drep_extract.R
\name{drep_extract}
\alias{drep_extract}
\title{Extract dRep genome results}
\usage{
drep_extract(path)
}
\arguments{
\item{path}{Path to the top level dRep directory}
}
\description{
Extract dRep genome results
}
