% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateRichnessFull.R
\name{estimate_richness_full}
\alias{estimate_richness_full}
\title{Alpha-Diversity Measures}
\usage{
estimate_richness_full(phyloObject, runPD = FALSE)
}
\description{
An extension of the phyloseq `estimate_richness` function to include Faith's PD and Shannon's Evenness
}
