% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_named_palette.R
\name{expand_named_palette}
\alias{expand_named_palette}
\title{Expanded Color Palette with Names}
\usage{
expand_named_palette(pal, return_list = TRUE)
}
\arguments{
\item{pal}{A (named) vector with colors}

\item{return_list}{Whether to return the values as a list (TRUE, default) or a named vector (FALSE).}
}
\description{
Give a vector or a named vector of rgb hex colors, and this will expand them
out adding alpha values, and appending 1, 2, 3, etc, to the names. If there
are no names, then they will be assigned names from the alphabet.
}
