% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_filter_summary.R
\name{fastq_filter_summary}
\alias{fastq_filter_summary}
\title{Summarize a fastq_filter.py file}
\usage{
fastq_filter_summary(file, plot = TRUE, md = TRUE)
}
\arguments{
\item{file}{Output from fastq_filter.py}

\item{plot}{Return a ggplot object}

\item{md}{Return tables in markdown format}
}
\description{
Summarize a fastq_filter.py file
}
