% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_summary.R
\name{fastq_info_summary}
\alias{fastq_info_summary}
\title{Summarize a fastq_info file}
\usage{
fastq_info_summary(file)
}
\arguments{
\item{file}{Output from fastq_info.py}
}
\description{
Summarize a fastq_info file
}
