% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dt.R
\name{format_dt}
\alias{format_dt}
\title{Format DT Datatable}
\usage{
format_dt(x, row_count = 10, rownames = FALSE, filename = "file")
}
\arguments{
\item{row_count}{Number of rows in the default view}

\item{rownames}{Include rownames}
}
\description{
Format DT Datatable
}
