% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fts_intervals.R
\name{get_fts_intervals}
\alias{get_fts_intervals}
\title{Get intervals for a run_fts() dataframe}
\usage{
get_fts_intervals(df)
}
\arguments{
\item{df}{A dataframe from run_fts()}
}
\description{
Get intervals for a run_fts() dataframe
}
