% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phylo_palette.R
\name{get_phylo_palette}
\alias{get_phylo_palette}
\title{getPhyloPalette}
\usage{
get_phylo_palette(p, n = 10)
}
\arguments{
\item{p}{A Phyloseq object}

\item{n}{Max number of taxa to be included in the "primary" palette, with the rest in a secondary, less dramatic palette.}
}
\description{
getPhyloPalette
}
