% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhyloDiversity.R
\name{getPhyloDiversity}
\alias{getPhyloDiversity}
\title{Phylogenetic Diversity Measures}
\usage{
getPhyloDiversity(p, runs = 999, null.model = "frequency")
}
\arguments{
\item{p}{A phyloseq object}

\item{runs}{Number of bootstraps to run}

\item{null.model}{The null.model for the SES calculations}
}
\description{
Phylogenetic Diversity Measures
}
