% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhyloPalette.R
\name{getPhyloPalette}
\alias{getPhyloPalette}
\title{getPhyloPalette}
\usage{
getPhyloPalette(p, level = "phylum", n = 10)
}
\arguments{
\item{p}{A Phyloseq object}

\item{level}{The taxonomic level to make a palette for}

\item{n}{Max number of taxa to be included in the "primary" palette, with the rest in a secondary, less dramatic palette.}
}
\description{
getPhyloPalette
}
