% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSig.R
\name{getSig}
\alias{getSig}
\title{getSig}
\usage{
getSig(group = "group", s = sigtab, phylo = phyloTemp)
}
\description{
Produces a df and plot of the significant taxa in a DESeq sigtab. Requires the phyloseq object and contrast field.
}
