% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helper.R
\name{gt_theme}
\alias{gt_theme}
\title{Web theme for GT}
\usage{
gt_theme(
  data,
  align = "center",
  background_color = "#222222",
  font_color = "#DDDDDD",
  column_label_size = 18,
  table_font_size = 16,
  table_padding = 2,
  ...
)
}
\arguments{
\item{data}{gt object}

\item{align}{column alignment (left, right, center)}

\item{background_color}{table background}

\item{font_color}{font color}

\item{column_label_size}{header font size}

\item{table_font_size}{table font size}
}
\description{
Web theme for GT
}
