% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illumina_codes.R
\name{illumina_codes}
\alias{illumina_codes}
\title{Get Illumina machine and flow cell code from run id}
\usage{
illumina_codes(run_id)
}
\arguments{
\item{run_id}{character string with the run id}
}
\description{
Machine codes and regular expressions scraped from users at
biostars (https://www.biostars.org/p/198143/).
}
