% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jak_theme_transparent.R
\name{jak_theme_transparent}
\alias{jak_theme_transparent}
\title{JAK transparent ggplot theme}
\usage{
jak_theme_transparent(base_size = 10, base_family = "",
  keySize = 0.3)
}
\description{
Produces a transparent background and light colors for figures with a dark background.
}
