% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthofinder.R
\name{orthofinder_complete_clusters}
\alias{orthofinder_complete_clusters}
\title{Summarize single copy and orthogroups found in all}
\usage{
orthofinder_complete_clusters(o)
}
\arguments{
\item{tsv_path}{The path to the Orthogroups/Orthogroups.tsv file in the standard orthofinder output}
}
\description{
Summarize single copy and orthogroups found in all
}
