% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthofinder.R
\name{orthofinder_summary}
\alias{orthofinder_summary}
\title{Summarize an orthofinder TSV file}
\usage{
orthofinder_summary(tsv_path)
}
\arguments{
\item{tsv_path}{The path to the Orthogroups/Orthogroups.tsv file in the standard orthofinder output}
}
\description{
Summarize an orthofinder TSV file
}
