% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.adonis.R
\name{pairwise.adonis}
\alias{pairwise.adonis}
\title{Calculate Total and Pairwise Adonis Values}
\usage{
pairwise.adonis(x, factors, sim.method = "bray", p.adjust.m = "BH",
  perm = 999)
}
\arguments{
\item{x}{A dataframe of values with samples in columns, and observations in rows}

\item{factors}{A dataframe with samples to be compared as rownames, and the groupings as column 1. It is easy to get this with a "select()" statement.}

\item{sim.method}{The distance method}

\item{p.adjust.m}{P-value adjustment method}

\item{perm}{= Number of permutations}
}
\description{
Calculate Total and Pairwise Adonis Values
}
