% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jakPalette.R
\docType{data}
\name{paletteJAK}
\alias{paletteJAK}
\title{A Color Palette Function}
\format{
An object of class \code{list} of length 2.
}
\usage{
paletteJAK
}
\description{
This palette can be used all over
}
\examples{
jakPalette(n)
}
\keyword{datasets}
