% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{permanova}
\alias{permanova}
\title{Permanova (adonis) and visualization}
\usage{
permanova(phyloObject, group = "TYPE", metric = "jsd",
  palettePick = FALSE, ordPoints = "none")
}
\description{
Permanova (adonis) and visualization
}
