% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_other.R
\name{phylo_other}
\alias{phylo_other}
\title{Combine low-abundant taxa into a single category}
\usage{
phylo_other(p, level = "phylum", n = 10)
}
\arguments{
\item{p}{A phyloseq object}

\item{level}{Taxonomic level to use}

\item{n}{Number of taxa to show, in addition to "Other"}
}
\description{
Combine low-abundant taxa into a single category
}
