% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloOther.R
\name{phyloOther}
\alias{phyloOther}
\title{Combine low-abundant taxa into a single category}
\usage{
phyloOther(p, level = "family", min = 10, xAxis = "SAMPLE",
  otherColor = "#BBBBBB")
}
\arguments{
\item{p}{A phyloseq object}

\item{level}{Taxonomic level to use}

\item{min}{Number of taxa to show, in addition to "_Other"}

\item{xAxis}{Variable for the x axis}

\item{otherColor}{Color (rgb) to assign to the "_Other" grouping}
}
\description{
Combine low-abundant taxa into a single category
}
