% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fts_intervals.R
\name{plot_fts_intervals}
\alias{plot_fts_intervals}
\title{Plot intervals for a run_fts() dataframe}
\usage{
plot_fts_intervals(df, n = 10, feature = c(), p = colorbook$modernUI(2))
}
\arguments{
\item{df}{A dataframe from run_fts()}

\item{n}{Subset by the top n features as determined by the order of the features in df}

\item{feature}{A feature (or vector of features) to subset. Overrides the n parameter}
}
\description{
Plot intervals for a run_fts() dataframe
}
