% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPD.R
\name{plotPD}
\alias{plotPD}
\title{Plot Faith's PD}
\usage{
plotPD(phyloObject)
}
\description{
This function iteratively removes the least abundant sequence in a phyloseq sample, re-roots the tree, calculates Faith's PD, and plots the results.
Returns a list with a dataframe (results$df) and two different plots.
}
\details{
results$p1 is from iteratively removing the lowest abundance sequences per sample. This is done on a per sample basis - if
a sequence is lowest in sample A but not sample B, it is removed from sample A but retained in sample B until it is lowest there.
If there are multiple sequences tied for lowest in a sample, they are all removed at once - hence there can be jumps along the x axis.

results$p2 uses the same data and criteria as results$p1, except now the x axis is the sequence count when removed. An X value of 10 for example
means that is the PD with all sequences of abundance 10 or less removed.
}
