% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqtab_stats.R
\name{seqtab_stats}
\alias{seqtab_stats}
\title{Extract DADA2 seqtab statistics}
\usage{
seqtab_stats(seqtab)
}
\arguments{
\item{seqtab}{DADA2 seqtab object}
}
\description{
Extract DADA2 seqtab statistics
}
