% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showColorPalette.R
\name{show_all_color_palettes}
\alias{show_all_color_palettes}
\title{Show all color palettes in the jak_palettes object}
\usage{
show_all_color_palettes()
}
\description{
Show all color palettes in the jak_palettes object
}
