% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showColorPalette.R
\name{show_color_palette}
\alias{show_color_palette}
\title{showColorPalette}
\usage{
show_color_palette(pal, labels = FALSE, label_angle = 0)
}
\arguments{
\item{pal}{A color palette}
}
\description{
showColorPalette
}
