% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigCombine.R
\name{sigCombine}
\alias{sigCombine}
\title{Combine DESeq sigtab contrasts}
\usage{
sigCombine(group = "group", reference = reference,
  comparison = comparison, sigtab)
}
\description{
Combine DESeq sigtab contrasts
}
