% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySE.R
\name{summarySE}
\alias{summarySE}
\title{Calculate Error Bars on a Dataframe}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\description{
Code borrowed from Winston Chang
http://www.cookbook-r.com/Graphs/Plotting_means_and_error_bars_(ggplot2)/
}
