% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolPalette.R
\docType{data}
\name{tol}
\alias{tol}
\title{Paul Tol Color Schemes}
\format{An object of class \code{list} of length 10.}
\usage{
tol
}
\description{
https://personal.sron.nl/~pault/data/colourschemes.pdf
}
\keyword{datasets}
