% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_summary.R
\name{fastq_info_summary}
\alias{fastq_info_summary}
\title{Summarize a fastq_info.py file}
\usage{
fastq_info_summary(file, plot = TRUE, md = TRUE)
}
\arguments{
\item{file}{Output from fastq_info.py}

\item{plot}{Return a ggplot object}

\item{md}{Return tables in markdown format}
}
\description{
Summarize a fastq_info.py file
}
