% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phylo_palette.R
\name{get_phylo_palette}
\alias{get_phylo_palette}
\title{getPhyloPalette}
\usage{
get_phylo_palette(
  p,
  n = 10,
  palette_high = colorbook$helpcenter,
  palette_low = colorbook$inslife,
  random_high = FALSE,
  random_low = FALSE,
  other_color = "#555555",
  OTU = F
)
}
\arguments{
\item{p}{A Phyloseq object}

\item{n}{Max number of taxa to be included in the "primary" palette, with the rest in a secondary, less dramatic palette}

\item{palette_high}{Expanding color palette for the high taxa}

\item{palette_low}{Expanding color palette for the low taxa}

\item{random_high}{If TRUE, randomize the color palette. If FALSE, colors will be arranged from high to low taxa abundance}

\item{random_low}{If TRUE, randomize the color palette. If FALSE, colors will be arranged from high to low taxa abundance}

\item{other_color}{Color to use for the aggregated Other taxa}
}
\description{
getPhyloPalette
}
