% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlphaDiversity.R
\name{getAlphaDiversity}
\alias{getAlphaDiversity}
\title{Alpha-Diversity Measures}
\usage{
getAlphaDiversity(p)
}
\arguments{
\item{p}{A phyloseq object}
}
\description{
An extension of the phyloseq `estimate_richness` function to include Shannon's
Evenness and sample depth. It also combines the normal estimate_richness output
with the sample data dataframe.
}
