% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhyloDiversity.R
\name{getPhyloDiversity}
\alias{getPhyloDiversity}
\title{Phylogenetic Diversity Measures (deprecated)}
\usage{
getPhyloDiversity(p)
}
\arguments{
\item{p}{A phyloseq object}

\item{runs}{Number of bootstraps to run}

\item{null.model}{The null.model for the SES calculations}

\item{minOccurrence}{Species must be found at least this many times in at least one sample to be kept}
}
\description{
Phylogenetic Diversity Measures (deprecated)
}
