% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jak_theme_black.R
\name{jak_theme_black}
\alias{jak_theme_black}
\title{JAK black ggplot theme}
\usage{
jak_theme_black(base_size = 10, base_family = "", keySize = 0.3)
}
\description{
Produces a black background and light colors for figures with a dark background.
}
