% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.adonis.R
\name{pairwise.adonis}
\alias{pairwise.adonis}
\title{Alpha-Diversity Measures}
\usage{
pairwise.adonis(x, factors, sim.method = "bray", p.adjust.m = "BH",
  perm = 999)
}
\arguments{
\item{x}{A dataframe of values with samples in columns, and observations in rows}

\item{factors}{A dataframe with samples to be compared as rownames, and the groupings as column 1. It is easy to get this with a "select()" statement.}

\item{sim.method}{The distance method}

\item{p.adjust.m}{P-value adjustment method}

\item{perm}{= Number of permutations}
}
\description{
An extension of the phyloseq `estimate_richness` function to include Shannon's Evenness and Faith's PD (with runPD = TRUE). It also combines the normal estimate_richness output with the sample data dataframe.
Running Faith's PD can add a significant amount of time if the tree is large. Faith's PD also requires a rooted tree. If the phyloseq object tree is not rooted, then this function loads the phytools package and does midpoint.root().
}
