% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jak_palettes.R
\docType{data}
\name{palette_jak}
\alias{palette_jak}
\title{A Color Palette Function}
\format{
An object of class \code{list} of length 7.
}
\usage{
palette_jak
}
\arguments{
\item{n}{Number of colors to return}
}
\description{
Miscellaneous palettes
}
\details{
Bay, winter, lake and sailboat from https://github.com/jakelawlor/PNWColors
}
\examples{
palette_jak$jak$blue1
}
\keyword{datasets}
