% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{permanova}
\alias{permanova}
\title{Permanova (adonis) and visualization}
\usage{
permanova(phyloObject, group = "TYPE", metric = "jsd",
  ordPoints = "none")
}
\arguments{
\item{phyloObject}{A phyloseq object}

\item{group}{A column name from the phyloseq object sample data to group samples for comparison}

\item{metric}{The distance metric to use}

\item{ordPoints}{X,Y points to use. If none given, then an NMDS will be run}

\item{palettePick}{A palette object if you don't want to use the default}
}
\description{
Permanova (adonis) and visualization
}
