% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistances.R
\name{plotDistances}
\alias{plotDistances}
\title{plotDistances() (deprecated)}
\usage{
plotDistances(
  p = GlobalPatterns,
  m = "wunifrac",
  s = "X.SampleID",
  d = "SampleType",
  plot = TRUE
)
}
\description{
plotDistances() (deprecated)
}
