% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistances.R
\name{plotDistances}
\alias{plotDistances}
\title{Boxplot of Beta-diversity Distances}
\usage{
plotDistances(p = GlobalPatterns, m = "wunifrac", s = "X.SampleID",
  d = "SampleType")
}
\description{
Takes a `phyloseq` object with samples grouped in the sample data, calculates all pairwise beta-diversity metrics, and plots the distances in facets.
}
