% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigCombine.R
\name{sigCombine}
\alias{sigCombine}
\title{Combine DESeq sigtab contrasts}
\usage{
sigCombine(group = "group", dds = dds, reference = reference,
  comparison = comparison, sigtab, runBH = TRUE)
}
\arguments{
\item{group}{Grouping used in the dds object}

\item{reference}{The name of the reference group}

\item{comparison}{The name of the comparison group}

\item{sigtab}{The sigtab object}

\item{runBH}{P-value corrections using BH}
}
\description{
Combine DESeq sigtab contrasts
}
